code/code/.ipynb_checkpoints/requirements-checkpoint.txt
pytorch-crf==0.7.2
transformers==4.18.0
datasets
seqeval